+++ Changelog.md
@@ -2,16 +2,29 @@
 
 ### Ignition Gazebo 2.X.X
 
+1. Support conversion and serialization of scene and light components
+    * [Pull Request 297](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/297)
+
 1. Use scene ambient and background color information in sensor
-   configuration. 
+   configuration.
     * [Pull Request 268](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/268)
 
 1. Added an SDF message to the start of log files.
     * [Pull Request 257](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/257)
+    
+1. Update Magnetometer component to use sdf::Sensor object instead of an sdf::ElementPtr.
+    * [Pull Request 272](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/272)
+
+1. Update Altimeter component to use sdf::Sensor object instead of an
+   sdf::ElementPtr.
+    * [Pull Request 286](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/286)
 
 ## Ignition Gazebo 1.x
 
 ### Ignition Gazebo 1.X.X
+
+1. Add Wind Plugin (Ported from Gazebo classic)
+    * [Pull Request 273](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/273/)
 
 1. Add `Link`: a convenience class for interfacing with link entities
     * [Pull Request 269](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/269)